// LifeInsuranceUI/src/app/models/employee-application.ts
export class EmployeeApplication {
    Id: number;
    FirstName: string;
    LastName: string;
    Email: string;
    Gender: string;
    PhoneNumber: number;
    DateOfBirth: string;
    MartialStatus: string;
    AadharNumber: number;
    Address: string;
    CurrentOccupation: string;
    JobRole: string;
    Cgpa_X: string;
    Cgpa_XII: string;
    CollegeCourse: string;
    Cgpa_College: string;
}